fix xbox compile errors

+++ src/vm_vmem.c
@@ -48,8 +48,8 @@ static VmAllocation* VM_FindAllocForPointer(void* p) {
 A_NO_DISCARD void* VM_Alloc(size_t n, VmAllocType type) {
     void* p = Z_Alloc(n);
     assert(p);
-    assert(p != 0xDDDDDDDD);
-    assert(p != 0xDDDDDDE1);
+    assert((size_t)p != 0xDDDDDDDD);
+    assert((size_t)p != 0xDDDDDDE1);
     VmAllocation* a = VM_FindFreeAlloction();
     assert(a);
     a->p = p;
@@ -65,8 +65,8 @@ A_NO_DISCARD void* VM_Alloc(size_t n, VmAllocType type) {
 A_NO_DISCARD void* VM_Zalloc(size_t n, VmAllocType type) {
     void* p = Z_Zalloc(n);
     assert(p);
-    assert(p != 0xDDDDDDDD);
-    assert(p != 0xDDDDDDE1);
+    assert((size_t)p != 0xDDDDDDDD);
+    assert((size_t)p != 0xDDDDDDE1);
     VmAllocation* a = VM_FindFreeAlloction();
     assert(a);
     a->p = p;
@@ -95,8 +95,8 @@ A_NO_DISCARD void* VM_AllocAt(const void* p, size_t n, VmAllocType type) {
 }
 
 bool VM_Free(void* p, VmAllocType type) {
-    assert(p != 0xDDDDDDDD);
-    assert(p != 0xDDDDDDE1);
+    assert((size_t)p != 0xDDDDDDDD);
+    assert((size_t)p != 0xDDDDDDE1);
     if (p == NULL)
         return true;
 
@@ -122,8 +122,8 @@ bool VM_Free(void* p, VmAllocType type) {
 }
 
 bool VM_FreeAt(void* p, VmAllocType type) {
-    assert(p != 0xDDDDDDDD);
-    assert(p != 0xDDDDDDE1);
+    assert((size_t)p != 0xDDDDDDDD);
+    assert((size_t)p != 0xDDDDDDE1);
     VmAllocation* a = VM_FindAllocForPointer(p);
     assert(a);
     if (a == NULL) {
